pyproject.toml
[tool.poetry]
name = "presto-requests"
version = "1.0.0"
description = ""
authors = ["Phillip Sitbon <phillip.sitbon@gmail.com>"]
readme = "README.md"

packages = [
  { include = "presto" },
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.extras]
async = ["httpx"]

[tool.poetry.dependencies]
python = "^3.11.0rc1"
requests = "^2.28.1"
attrdict3 = "^2.0.2"
httpx = {version = "^0.23.1", optional = true}
